refactor: move database data to separate module
